+++ src/pymovements/dataset/_utils/_archives.py
@@ -35,6 +35,7 @@ from typing import IO
 from tqdm import tqdm
 
 from pymovements._utils._paths import get_filepaths
+from pymovements.exceptions import UnknownFileType
 
 
 def extract_archive(
@@ -269,11 +270,11 @@ def _detect_file_type(filepath: Path) -> tuple[str | None, str | None]:
 
     Raises
     ------
-    RuntimeError
+    UnknownFileType
         If the file has no suffix or the suffix is not supported.
     """
     if not (suffixes := filepath.suffixes):
-        raise RuntimeError(
+        raise UnknownFileType(
             f"File '{filepath}' has no suffixes that could be used to detect the archive type or"
             ' compression.',
         )
@@ -296,9 +297,9 @@ def _detect_file_type(filepath: Path) -> tuple[str | None, str | None]:
 
             # Check if the second last suffix refers to an archive type.
             if (suffix2 := suffixes[-2]) not in _ARCHIVE_EXTRACTORS:
-                raise RuntimeError(
-                    f"Unsupported archive type: '{suffix2}'.\n"
-                    f"Supported suffixes are: '{sorted(set(_ARCHIVE_EXTRACTORS))}'.",
+                raise UnknownFileType(
+                    f"Unsupported compression or archive type: '{suffix2}{suffix}'.\n"
+                    f"Supported suffixes are: '{_get_valid_suffixes()}'.",
                 )
             # We detected a compressed archive file (e.g. tar.gz).
             return suffix2, suffix
@@ -307,12 +308,9 @@ def _detect_file_type(filepath: Path) -> tuple[str | None, str | None]:
         return None, suffix
 
     # Raise error as we didn't find a valid suffix.
-    valid_suffixes = sorted(
-        set(_ARCHIVE_TYPE_ALIASES) | set(_ARCHIVE_EXTRACTORS) | set(_COMPRESSED_FILE_OPENERS),
-    )
-    raise RuntimeError(
+    raise UnknownFileType(
         f"Unsupported compression or archive type: '{suffix}'.\n"
-        f"Supported suffixes are: '{valid_suffixes}'.",
+        f"Supported suffixes are: '{_get_valid_suffixes()}'.",
     )
 
 
@@ -362,3 +360,10 @@ def _decompress(
         source_path.unlink()
 
     return destination_path
+
+
+def _get_valid_suffixes() -> list[str]:
+    """Get valid archive file extensions."""
+    return sorted(
+        set(_ARCHIVE_TYPE_ALIASES) | set(_ARCHIVE_EXTRACTORS) | set(_COMPRESSED_FILE_OPENERS),
+    )
