+++ src/pymovements/exceptions.py
@@ -36,3 +36,7 @@ class InvalidProperty(Exception):
     def __init__(self, property_name: str, valid_properties: list[str]):
         message = f"property '{property_name}' is invalid. Valid properties are: {valid_properties}"
         super().__init__(message)
+
+
+class UnknownFileType(RuntimeError):
+    """Raised on unknown file types."""
