+++ tests/unit/dataset/_utils/_downloads_test.py
@@ -26,7 +26,6 @@ import pytest
 
 from pymovements.dataset._utils._downloads import _DownloadProgressBar
 from pymovements.dataset._utils._downloads import _get_redirected_url
-from pymovements.dataset._utils._downloads import download_and_extract_archive
 from pymovements.dataset._utils._downloads import download_file
 
 
@@ -148,88 +147,6 @@ def test__get_redirected_url_with_redirects_max_hops():
         'https://github.com/aeye-lab/pymovements/archive/master.zip '\
         'exceeded 0 redirects. The last redirect points to '\
         'https://codeload.github.com/aeye-lab/pymovements/zip/main.'
-
-
-def test_download_and_extract_archive(tmp_path):
-    url = 'https://github.com/aeye-lab/pymovements/archive/refs/tags/v0.4.0.tar.gz'
-    download_filename = 'pymovements-0.4.0.tar.gz'
-    md5 = '52bbf03a7c50ee7152ccb9d357c2bb30'
-    extract_dirpath = tmp_path / 'extracted'
-
-    download_and_extract_archive(
-        url,
-        tmp_path,
-        download_filename,
-        extract_dirpath,
-        md5,
-        remove_top_level=False,
-    )
-
-    assert extract_dirpath.exists()
-    assert (extract_dirpath / 'pymovements-0.4.0').exists()
-    assert (extract_dirpath / 'pymovements-0.4.0' / 'README.md').exists()
-
-    archive_path = tmp_path / download_filename
-    assert archive_path.exists()
-    with open(archive_path, 'rb') as f:
-        file_bytes = f.read()
-        assert hashlib.md5(file_bytes).hexdigest() == md5
-
-
-@pytest.mark.parametrize(
-    'verbose',
-    [
-        pytest.param(False, id='verbose_false'),
-        pytest.param(True, id='verbose_true'),
-    ],
-)
-def test_download_and_extract_archive_extract_dirpath_None(tmp_path, capsys, verbose):
-    url = 'https://github.com/aeye-lab/pymovements/archive/refs/tags/v0.4.0.tar.gz'
-    download_filename = 'pymovements-0.4.0.tar.gz'
-    md5 = '52bbf03a7c50ee7152ccb9d357c2bb30'
-    extract_dirpath = None
-
-    # extract first time
-    download_and_extract_archive(
-        url, tmp_path, download_filename, extract_dirpath, md5, verbose=verbose,
-    )
-    out, _ = capsys.readouterr()
-
-    if verbose:
-        assert out == 'Downloading '\
-            'https://github.com/aeye-lab/pymovements/archive/refs/tags/v0.4.0.tar.gz'\
-            f" to {os.path.join(tmp_path, 'pymovements-0.4.0.tar.gz')}\n" \
-            f'Checking integrity of pymovements-0.4.0.tar.gz\n'\
-            f'Extracting pymovements-0.4.0.tar.gz to {tmp_path}\n'
-    else:
-        assert out == ''
-
-    # extract second time to test already downloaded and verified file
-    download_and_extract_archive(
-        url, tmp_path, download_filename, extract_dirpath, md5, verbose=verbose,
-    )
-    out, _ = capsys.readouterr()
-
-    if verbose:
-        assert out == f'Using already downloaded and verified file: '\
-            f"{os.path.join(tmp_path, 'pymovements-0.4.0.tar.gz')}"\
-            f'\nExtracting pymovements-0.4.0.tar.gz to {tmp_path}\n'
-    else:
-        assert out == ''
-
-
-def test_download_and_extract_archive_invalid_md5(tmp_path):
-    url = 'https://github.com/aeye-lab/pymovements/archive/refs/tags/v0.4.0.tar.gz'
-    download_filename = 'pymovements-0.4.0.tar.gz'
-    md5 = '00000000000000000000000000000000'
-    extract_dirpath = tmp_path / 'extracted'
-
-    with pytest.raises(RuntimeError) as excinfo:
-        download_and_extract_archive(url, tmp_path, download_filename, extract_dirpath, md5)
-
-    msg, = excinfo.value.args
-    assert msg == f"File {os.path.join(tmp_path, 'pymovements-0.4.0.tar.gz')} "\
-        'not found or download corrupted.'
 
 
 def test__DownloadProgressBar_tsize_not_None():
